requirements.txt
Django>=5.2
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
dj-database-url
python-dotenv
